+++ js/form/framelix-form-field-password.js
@@ -7,7 +7,7 @@ class FramelixFormFieldPassword extends FramelixFormFieldText {
    * Maximal width in pixel
    * @type {number|null}
    */
-  maxWidth = 300
+  maxWidth = 400
 
   /**
    * Render the field into the container
@@ -17,7 +17,7 @@ class FramelixFormFieldPassword extends FramelixFormFieldText {
     await super.renderInternal()
     const self = this
     this.field.attr('data-field-with-button', '1')
-    this.field.append(`<button class="framelix-button" title="__framelix_form_password_toggle__" type="button"><span class="material-icons">visibility</span></button>`)
+    this.field.append(`<button class="framelix-button framelix-button-primary" title="__framelix_form_password_toggle__" type="button" data-icon-left="visibility"></button>`)
     this.field.find('.framelix-button').on('click keydown', function (ev) {
       if (ev.key === ' ' || ev.key === 'Enter' || !ev.key) {
         self.input.attr('type', self.input.attr('type') === self.type ? 'text' : 'password')